+++ app/helpers/backoffice/offers_helper.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Backoffice::OffersHelper
+  def save_as_draft_title
+    "Save as draft"
+  end
+
+  def exit_title
+    "Exit without saving"
+  end
+end