+++ config/routes.rb
@@ -138,6 +138,10 @@ Rails.application.routes.draw do
   post "/backoffice/services/:service_id/offers/:offer_id/submit", to: "backoffice/services/offers#submit_summary"
   patch "/backoffice/services/:service_id/offers/:offer_id/submit", to: "backoffice/services/offers#submit_summary"
 
+  post "/backoffice/services/:service_id/offers/:offer_id/duplicate", to: "backoffice/services/offers#duplicate", 
+as: :duplicate_offer
+  
+
   resource :executive, only: :show
   namespace :executive do
     resources :statistics, only: :index